kas/requirements.txt
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
openai>=1.0.0
